# NaN threshold for conservative method