--- a/src/xarray_regrid/utils.py
+++ b/src/xarray_regrid/utils.py
@@ -1,5 +1,5 @@
+from collections.abc import Callable
 from dataclasses import dataclass
-from typing import Callable
 
 import numpy as np
 import pandas as pd