Nice package! In testing it out where I've previously used `xesmf`, I noticed two features lacking from the conservative method:

1. the regridded dataset can't be constructed lazily if dask-backed due to [this line](https://github.com/EXCITED-CO2/xarray-regrid/blob/main/src/xarray_regrid/methods/conservative.py#L224)
2. no ability to keep target cells where the input points are partially NaN, as noted in #32

Number 1 is easy, number 2 is trickier. I added a naive implementation for the `nan_threshold` capabilities of xesmf here for discussion. As noted in the previous issue, to do this 100% correctly we would need to track the NaN fraction as we reduce over each dimension, which I'm not doing here. The `nan_threshold` value doesn't translate directly to total fraction of NaN cells due to the sequential reduction. It would also get complicated for isolated NaNs in the temporal dimension.

I'm not sure any of this matters much for a dataset where you have consistent NaN's e.g. SST. Here's an example of the new functionality used on the [MUR dataset](https://registry.opendata.aws/mur/). Note this is a 33TB array but we can now generate the (lazily) regridded dataset instantaneously.

```python
import xarray as xr
import xarray_regrid

sst = xr.open_zarr("https://mur-sst.s3.us-west-2.amazonaws.com/zarr-v1").analysed_sst
grid = xarray_regrid.Grid(
    north=90,
    east=180,
    south=-90,
    west=-180,
    resolution_lat=1,
    resolution_lon=1,
)
target = xarray_regrid.create_regridding_dataset(grid)
target = target.rename(latitude="lat", longitude="lon")

ds0p0 = sst.regrid.conservative(target, nan_threshold=0)
ds0p5 = sst.regrid.conservative(target, nan_threshold=0.5)
ds1p0 = sst.regrid.conservative(target, nan_threshold=1)
```
![ds0p0](https://github.com/EXCITED-CO2/xarray-regrid/assets/39069044/e5b47106-62d7-4eb1-bad5-a764462529c8)

![ds0p5](https://github.com/EXCITED-CO2/xarray-regrid/assets/39069044/2fd07b38-040a-4a8b-b89b-fd986e365419)

![ds1p0](https://github.com/EXCITED-CO2/xarray-regrid/assets/39069044/0141e11e-c2d9-45d2-bc2d-0542d98b0943)