@slevang 2024-05-26
Thanks, and yes for sure, hopefully next week I will find time to add tests and benchmarks.