@slevang 2024-05-26
Did some quick profiling on a ~4GB array of 1/4deg global data coarsening to 1deg. Dask array on a 32 CPU node. Results:

- This PR, `skipna=False`: 32s
- This PR, `skipna=True`: 64s
- `main`: 96s

So adding skipna forces roughly one additional pass through the array with the weight renormalization. The reason this PR is faster than `main` is because the current code has the `np.any(np.isnan())` check which forces computation, plus the separately calculated `isnan` array, which forces 3 passes through the data. If I cut out the logic branch of checking for NaNs on `main` and go straight to the einsum, we recover the ~32s run above.