@slevang 2024-07-14
Made the modification to take `notnull.any(non_regrid_dims)` which leaves us at about a 3x performance penalty for `skipna=True` in the benchmarks I've run. I think this should maybe be a configurable arg though in cases where you want to track NaNs very carefully throughout the dataset.